refactor: merge v2 components back to main path
fix: delete v2 folder
