+++ operate/client/src/App/Dashboard/IncidentsByError/index.test.tsx
@@ -17,20 +17,26 @@ import {
 import {IncidentsByError} from './index';
 import {
   mockIncidentsByError,
-  mockIncidentsByErrorWithBigErrorMessage,
+  mockIncidentStatisticsByErrorWithBigMessage,
+  mockIncidentStatisticsByDefinition,
   bigErrorMessage,
 } from './index.setup';
 import {LocationLog} from 'modules/utils/LocationLog';
-import {mockFetchIncidentsByError} from 'modules/mocks/api/incidents/fetchIncidentsByError';
 import {Paths} from 'modules/Routes';
-import {incidentsByErrorStore} from 'modules/stores/incidentsByError';
+import {mockFetchIncidentProcessInstanceStatisticsByError} from 'modules/mocks/api/v2/incidents/fetchIncidentProcessInstanceStatisticsByError';
+import {mockFetchIncidentProcessInstanceStatisticsByDefinition} from 'modules/mocks/api/v2/incidents/fetchIncidentProcessInstanceStatisticsByDefinition';
 import {QueryClientProvider} from '@tanstack/react-query';
 import {getMockQueryClient} from 'modules/react-query/mockQueryClient';
-import {createUser} from 'modules/testUtils';
+import {createUser, searchResult} from 'modules/testUtils';
 import {mockMe} from 'modules/mocks/api/v2/me';
+import {useEffect} from 'react';
+import {panelStatesStore} from 'modules/stores/panelStates';
+import {truncateErrorMessage} from './utils/truncateErrorMessage';
 
 function createWrapper(initialPath: string = Paths.dashboard()) {
   const Wrapper: React.FC<{children?: React.ReactNode}> = ({children}) => {
+    useEffect(() => () => panelStatesStore.reset(), []);
+
     return (
       <QueryClientProvider client={getMockQueryClient()}>
         <MemoryRouter initialEntries={[initialPath]}>
@@ -52,8 +58,23 @@ describe('IncidentsByError', () => {
     mockMe().withSuccess(createUser());
   });
 
+  const mockIncidentQueries = ({
+    incidents = mockIncidentsByError,
+    definitions = mockIncidentStatisticsByDefinition,
+  } = {}) => {
+    mockFetchIncidentProcessInstanceStatisticsByError().withSuccess(incidents);
+    mockFetchIncidentProcessInstanceStatisticsByError().withSuccess(incidents);
+
+    mockFetchIncidentProcessInstanceStatisticsByDefinition().withSuccess(
+      definitions,
+    );
+    mockFetchIncidentProcessInstanceStatisticsByDefinition().withSuccess(
+      definitions,
+    );
+  };
+
   it('should display skeleton when loading', async () => {
-    mockFetchIncidentsByError().withSuccess(mockIncidentsByError);
+    mockIncidentQueries();
 
     render(<IncidentsByError />, {
       wrapper: createWrapper(),
@@ -61,13 +82,15 @@ describe('IncidentsByError', () => {
 
     expect(screen.getByTestId('data-table-skeleton')).toBeInTheDocument();
 
-    await waitForElementToBeRemoved(
+    await waitForElementToBeRemoved(() =>
       screen.queryByTestId('data-table-skeleton'),
     );
+
+    expect(await screen.findByTestId('incident-byError')).toBeInTheDocument();
   });
 
   it('should handle server errors', async () => {
-    mockFetchIncidentsByError().withServerError();
+    mockFetchIncidentProcessInstanceStatisticsByError().withServerError();
 
     render(<IncidentsByError />, {
       wrapper: createWrapper(),
@@ -83,7 +106,7 @@ describe('IncidentsByError', () => {
       .spyOn(global.console, 'error')
       .mockImplementation(() => {});
 
-    mockFetchIncidentsByError().withNetworkError();
+    mockFetchIncidentProcessInstanceStatisticsByError().withNetworkError();
 
     render(<IncidentsByError />, {
       wrapper: createWrapper(),
@@ -96,8 +119,8 @@ describe('IncidentsByError', () => {
     consoleErrorMock.mockRestore();
   });
 
-  it('should display information message when there are no processes', async () => {
-    mockFetchIncidentsByError().withSuccess([]);
+  it('should display information message when there are no incidents', async () => {
+    mockIncidentQueries({incidents: searchResult([])});
 
     render(<IncidentsByError />, {
       wrapper: createWrapper(),
@@ -111,8 +134,8 @@ describe('IncidentsByError', () => {
     ).toBeInTheDocument();
   });
 
-  it('should render process incidents by error message', async () => {
-    mockFetchIncidentsByError().withSuccess(mockIncidentsByError);
+  it('should render process incidents with expandable process list', async () => {
+    mockIncidentQueries();
 
     const {user} = render(<IncidentsByError />, {
       wrapper: createWrapper(),
@@ -130,75 +153,71 @@ describe('IncidentsByError', () => {
 
     const processLink = withinIncident.getByRole('link', {
       description:
-        "View 36 Instances with error JSON path '$.paid' has no result.",
+        "View 78 Instances with error JSON path '$.paid' has no result.",
     });
 
     expect(processLink).toHaveAttribute(
       'href',
-      '/processes?errorMessage=JSON+path+%27%24.paid%27+has+no+result.&incidentErrorHashCode=234254&incidents=true',
+      `${Paths.processes()}?errorMessage=JSON+path+%27%24.paid%27+has+no+result.&incidentErrorHashCode=234254&incidents=true`,
     );
 
-    // this button click has no effect (check useEffect in Collapse component)
     await user.click(expandButton);
 
-    const firstVersion = await screen.findByRole('link', {
+    const firstProcess = await screen.findByRole('link', {
       description:
-        "View 37 Instances with error JSON path '$.paid' has no result. in version 1 of Process mockProcess",
+        "View 52 Instances with error JSON path '$.paid' has no result. in version 1 of Process Call Level 2 Process – Version 1",
     });
-    expect(
-      within(firstVersion).getByTestId('incident-instances-badge'),
-    ).toHaveTextContent('37');
-    expect(
-      within(firstVersion).getByText('mockProcess – Version 1'),
-    ).toBeInTheDocument();
 
-    expect(firstVersion).toHaveAttribute(
+    expect(
+      within(firstProcess).getByTestId('incident-instances-badge'),
+    ).toHaveTextContent('52');
+    expect(firstProcess).toHaveAttribute(
       'href',
-      '/processes?process=mockProcess&version=1&errorMessage=JSON+path+%27%24.paid%27+has+no+result.&incidentErrorHashCode=234254&incidents=true',
+      `${Paths.processes()}?process=call-level-2-process&version=1&errorMessage=JSON+path+%27%24.paid%27+has+no+result.&incidentErrorHashCode=234254&incidents=true`,
     );
   });
 
-  it('should update after next poll', async () => {
-    mockFetchIncidentsByError().withSuccess(mockIncidentsByError);
+  it('should include tenant in link when multi-tenancy is enabled', async () => {
+    vi.stubGlobal('clientConfig', {multiTenancyEnabled: true});
 
-    render(<IncidentsByError />, {
-      wrapper: createWrapper(),
-    });
-
-    const withinIncident = within(
-      await screen.findByTestId('incident-byError-0'),
-    );
-
-    expect(
-      withinIncident.getByRole('button', {
-        name: 'Expand current row',
-      }),
-    ).toBeInTheDocument();
-
-    mockFetchIncidentsByError().withSuccess([
-      {...mockIncidentsByError[0]!, instancesWithErrorCount: 40},
-      mockIncidentsByError[1]!,
-    ]);
-
-    expect(
-      await withinIncident.findByRole('button', {
-        name: 'Expand current row',
-      }),
-    ).toBeInTheDocument();
-
-    await waitFor(() =>
-      expect(incidentsByErrorStore.isPollRequestRunning).toBe(false),
-    );
-  });
-
-  it('should truncate the error message search param', async () => {
-    mockFetchIncidentsByError().withSuccess(
-      mockIncidentsByErrorWithBigErrorMessage,
-    );
+    mockIncidentQueries();
 
     const {user} = render(<IncidentsByError />, {
       wrapper: createWrapper(),
     });
+
+    const expandButton = within(
+      await screen.findByTestId('incident-byError-0'),
+    ).getByRole('button', {name: 'Expand current row'});
+
+    await user.click(expandButton);
+
+    const tenantProcess = await screen.findByRole('link', {
+      description:
+        "View 26 Instances with error JSON path '$.paid' has no result. in version 1 of Process Process with elements incidents – Version 1 – tenant-a",
+    });
+
+    expect(tenantProcess).toHaveAttribute(
+      'href',
+      `${Paths.processes()}?process=process-elements-incidents&version=1&errorMessage=JSON+path+%27%24.paid%27+has+no+result.&incidentErrorHashCode=234254&incidents=true&tenant=tenant-a`,
+    );
+  });
+
+  it('should truncate the error message in search params', async () => {
+    mockIncidentQueries({
+      incidents: mockIncidentStatisticsByErrorWithBigMessage,
+    });
+
+    const {user} = render(<IncidentsByError />, {
+      wrapper: createWrapper(),
+    });
+
+    const truncated = truncateErrorMessage(bigErrorMessage);
+    const mainSearch = new URLSearchParams({
+      errorMessage: truncated,
+      incidentErrorHashCode: '234254',
+      incidents: 'true',
+    }).toString();
 
     const withinIncident = within(
       await screen.findByTestId('incident-byError-0'),
@@ -212,30 +231,36 @@ describe('IncidentsByError', () => {
       withinIncident.getByRole('link', {
         description: `View 36 Instances with error ${bigErrorMessage}`,
       }),
-    ).toHaveAttribute(
-      'href',
-      '/processes?errorMessage=Lorem+ipsum+dolor+sit+amet%2C+consectetur+adipiscing+elit%2C+sed+do+eiusmod+tempor+incididunt+ut+labore&incidentErrorHashCode=234254&incidents=true',
-    );
+    ).toHaveAttribute('href', `${Paths.processes()}?${mainSearch}`);
 
-    // this button click has no effect (check useEffect in Collapse component)
     await user.click(expandButton);
 
-    expect(
-      await screen.findByRole('link', {
-        description: `View 37 Instances with error ${bigErrorMessage} in version 1 of Process mockProcess`,
-      }),
-    ).toHaveAttribute(
+    const firstProcess = await screen.findByRole('link', {
+      description: `View 52 Instances with error ${bigErrorMessage} in version 1 of Process Call Level 2 Process – Version 1`,
+    });
+
+    const detailSearch = new URLSearchParams({
+      process: 'call-level-2-process',
+      version: '1',
+      errorMessage: truncated,
+      incidentErrorHashCode: '234254',
+      incidents: 'true',
+    }).toString();
+
+    expect(firstProcess).toHaveAttribute(
       'href',
-      '/processes?process=mockProcess&version=1&errorMessage=Lorem+ipsum+dolor+sit+amet%2C+consectetur+adipiscing+elit%2C+sed+do+eiusmod+tempor+incididunt+ut+labore&incidentErrorHashCode=234254&incidents=true',
+      `${Paths.processes()}?${detailSearch}`,
     );
   });
 
   it('should expand filters panel on click', async () => {
-    mockFetchIncidentsByError().withSuccess(mockIncidentsByError);
+    mockIncidentQueries();
 
     const {user} = render(<IncidentsByError />, {
       wrapper: createWrapper(),
     });
+
+    expect(panelStatesStore.state.isFiltersCollapsed).toBe(false);
 
     const withinIncident = within(
       await screen.findByTestId('incident-byError-0'),
@@ -244,19 +269,21 @@ describe('IncidentsByError', () => {
     const expandButton = withinIncident.getByRole('button', {
       name: 'Expand current row',
     });
-    expect(expandButton).toBeInTheDocument();
 
-    const processLink = withinIncident.getByRole('link', {
+    await user.click(expandButton);
+
+    const processLink = await screen.findByRole('link', {
       description:
-        "View 36 Instances with error JSON path '$.paid' has no result.",
+        "View 52 Instances with error JSON path '$.paid' has no result. in version 1 of Process Call Level 2 Process – Version 1",
     });
+
+    panelStatesStore.toggleFiltersPanel();
+    expect(panelStatesStore.state.isFiltersCollapsed).toBe(true);
 
     await user.click(processLink);
 
     await waitFor(() =>
-      expect(screen.getByTestId('search')).toHaveTextContent(
-        /^\?errorMessage=JSON\+path\+%27%24.paid%27\+has\+no\+result.&incidentErrorHashCode=234254&incidents=true$/,
-      ),
+      expect(panelStatesStore.state.isFiltersCollapsed).toBe(false),
     );
   });
 });
