+++ operate/client/src/App/Dashboard/MetricPanel/index.tsx
