+++ operate/client/src/App/Dashboard/MetricPanel/styled.ts
