+++ operate/client/src/App/Dashboard/index.test.tsx
@@ -11,8 +11,8 @@ import {render, screen} from 'modules/testing-library';
 import {MemoryRouter} from 'react-router-dom';
 import {PAGE_TITLE} from 'modules/constants';
 import {Dashboard} from './index';
-import {mockIncidentsByError} from './v2/IncidentsByError/index.setup';
-import {mockWithSingleVersion} from './v2/InstancesByProcessDefinition/index.setup';
+import {mockIncidentsByError} from './IncidentsByError/index.setup';
+import {mockWithSingleVersion} from './InstancesByProcessDefinition/index.setup';
 import {mockFetchIncidentProcessInstanceStatisticsByError} from 'modules/mocks/api/v2/incidents/fetchIncidentProcessInstanceStatisticsByError';
 import {mockFetchProcessDefinitionStatistics} from 'modules/mocks/api/v2/processDefinitions/fetchProcessDefinitionStatistics';
 import {QueryClientProvider} from '@tanstack/react-query';
