+++ operate/client/src/App/Dashboard/index.tsx
@@ -8,11 +8,11 @@
 
 import {useEffect} from 'react';
 import {VisuallyHiddenH1} from 'modules/components/VisuallyHiddenH1';
-import {MetricPanel} from './v2/MetricPanel';
+import {MetricPanel} from './MetricPanel';
 import {PAGE_TITLE} from 'modules/constants';
 import {Grid, ScrollableContent, Tile, TileTitle} from './styled';
-import {InstancesByProcessDefinition} from './v2/InstancesByProcessDefinition';
-import {IncidentsByError} from './v2/IncidentsByError';
+import {InstancesByProcessDefinition} from './InstancesByProcessDefinition';
+import {IncidentsByError} from './IncidentsByError';
 import {useProcessDefinitionStatistics} from 'modules/queries/processDefinitionStatistics/useProcessDefinitionStatistics';
 import {NoInstancesEmptyState} from './NoInstancesEmptyState';
