--- operate/client/src/App/Dashboard/v2/IncidentsByError/utils/getAccordionTitle.tsx
@@ -1,21 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-
-import pluralSuffix from 'modules/utils/pluralSuffix';
-
-function getAccordionTitle(
-  instancesWithErrorCount: number,
-  errorMessage: string,
-) {
-  return `View ${pluralSuffix(
-    instancesWithErrorCount,
-    'Instance',
-  )} with error ${errorMessage}`;
-}
-
-export {getAccordionTitle};
