+++ operate/client/src/App/index.tsx
@@ -30,7 +30,6 @@ import {currentTheme} from 'modules/stores/currentTheme';
 import {ThemeSwitcher} from 'modules/components/ThemeSwitcher';
 import {ForbiddenPage} from 'modules/components/ForbiddenPage';
 import {ReactQueryProvider} from 'modules/react-query/ReactQueryProvider';
-import {getClientConfig} from '../modules/utils/getClientConfig';
 
 const Wrapper: React.FC = () => {
   return (
@@ -71,13 +70,8 @@ const routes = createRoutesFromElements(
       <Route
         index
         lazy={async () => {
-          if (getClientConfig()?.databaseType === 'rdbms') {
-            const {Dashboard} = await import('./Dashboard/v2/index');
-            return {Component: Dashboard};
-          } else {
-            const {Dashboard} = await import('./Dashboard/index');
-            return {Component: Dashboard};
-          }
+          const {Dashboard} = await import('./Dashboard/index');
+          return {Component: Dashboard};
         }}
       />
       <Route
